+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Update inquirer dependencies (#2501)
+- Improve endpoint type to be based on core type (#2511)
 
 ### Fixed
 - Not being able to select a custom repo when initializing a project (#2501)